% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\title{Analyze each variable and generate a histogram describing it's distribution.}
\usage{
distributions(data_analyze, outdir)
}
\arguments{
\item{data_analyze}{a data frame to analyze}

\item{outdir}{an optional output directory to save the resulting plots as png images}
}
\description{
Also returns a table of all numeric variables describind it's percentiles 1, 10, 25, 50 (median), 75, 90 and 99.
}
\examples{
library(xray)
distributions(mtcars)

}
