% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timebased.R
\name{timebased}
\alias{timebased}
\title{Analyze each variable in respect to a time variable}
\usage{
timebased(data_analyze, date_variable, time_unit = "auto", outdir)
}
\arguments{
\item{data_analyze}{a data frame to analyze}

\item{date_variable}{the variable that will be used to pivot all other variables}

\item{time_unit}{the time unit to use if not automatically}

\item{outdir}{an optional output directory to save the resulting plots as png images}
}
\description{
Analyze each variable in respect to a time variable
}
\examples{

data(longley)
longley$Year=as.Date(paste0(longley$Year,'-01-01'))
xray::timebased(longley, 'Year')

}
